+++ package.json
@@ -33,6 +33,7 @@
     "/directives",
     "/generated",
     "/helpers",
+    "!/helpers/internal",
     "/lang",
     "/mixins",
     "/templates",
